Using Parameter objects and adding rudimentary autoPilot support

+++ SparkleLayers/CloudLayer.pde
@@ -32,10 +32,10 @@ public class CloudLayer extends Layer {
         float m = this.fractalNoise(dx + x*scale, dy + y*scale, z + 10.0) - 0.75;
   
         color c = color(
-           ((hue + 80.0 * m) % 100.0),
-           (100 - 100 * constrain(pow(3.0 * n, 3.5), 0, 0.9)),
-           (100 * constrain(pow(3.0 * n, 1.5), 0, 0.9)),
-           map(a, 0.0, 1.0, 0, 127)
+            ((hue + 80.0 * m) % 100.0),
+            (100 - 100 * constrain(pow(3.0 * n, 3.5), 0, 0.9)),
+            (100 * constrain(pow(3.0 * n, 1.5), 0, 0.9)),
+            map(a, 0.0, 1.0, 0, 127)
            );
         
         this.pg.pixels[x + this.pg.width*y] = c;
@@ -63,8 +63,8 @@ public class CloudLayer extends Layer {
   }
 
   
-  public void setParam1(int v){}
-  public void setParam2(int v){}
-  public void setParam3(int v){}
+  public void setParam1(float v){}
+  public void setParam2(float v){}
+  public void setParam3(float v){}
 
 }
